tests/requirements.txt
# Testing Requirements
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-benchmark>=4.0.0
requests>=2.31.0
websockets>=11.0
psycopg2-binary>=2.9.6
redis>=5.0.0

